Add player interaction with height map

+++ src/window/Window.cpp
@@ -98,7 +98,7 @@ int Window::getHeight() const {
     return height;
 }
 
-double Window::getLastFrameTime() const {
+float Window::getLastFrameTime() const {
     return lastFrameTime;
 }
